- Refactor Raid Table - Create isolated columns
