+++ app-client/src/Page/Homepage/components/RaidTable/columns/endDateColumn.tsx
@@ -1,0 +1,11 @@
+import {GridColDef} from "@mui/x-data-grid";
+
+export const endDateColumn: GridColDef = {
+    field: "date.endDate",
+    headerName: "End Date",
+    width: 100,
+    renderCell: (params) => {
+        return params.row.date.endDate;
+    },
+    sortable: false,
+}
